Set `autorun: false` by default everywhere.
Previously the cell option `autorun` defaulted to `true` for
"sandbox" type cells, but `false` for exercises. This has been
found to be confusing, and so `autorun: false` is now always the
default state. Autorun may still be enabled by setting the cell
option directly, or set document-wide using the `cell-options`
YAML header.